+++ test/unit/encoder.spec.mjs
@@ -1,7 +1,7 @@
-import Encoder from '../src/encoder.mjs';
-import {ArgumentError} from '../src/errors.mjs';
+import Encoder from '../../src/encoder.mjs';
+import {ArgumentError} from '../../src/errors.mjs';
 import assert from 'assert';
-import Base32 from '../utils/base32.mjs';
+import Base32 from '../../utils/base32.mjs';
 
 describe('[Encoder] class tests', function () {
   describe('decode method', function () {
