+++ test/unit/resources/core.mjs
